UI Automator: Refactor `setUp()` and `tearDown()` to `UiAutomatorTest`
